NPM Start error fix
Previously, if you ran npm start following the directions, you were greeted with an error that said command was not ound. By adding this single line to package.json, it fixes the issue and makes damn sure the commands listed in the readme actually work. I would appreciate it if this could be made part of the parent repository. Thanks!